Assert byte counts for metrics (#975)

+++ com.unity.multiplayer.mlapi/Runtime/Logging/NetworkLog.cs
@@ -66,7 +66,7 @@ namespace MLAPI.Logging
                     writer.WriteStringPacked(message);
 
                     NetworkManager.Singleton.MessageSender.Send(NetworkManager.Singleton.ServerClientId, NetworkConstants.SERVER_LOG, NetworkChannel.Internal, buffer);
-                    NetworkManager.Singleton.NetworkMetrics.TrackServerLogSent(NetworkManager.Singleton.ServerClientId, (uint)logType, (ulong)buffer.Length);
+                    NetworkManager.Singleton.NetworkMetrics.TrackServerLogSent(NetworkManager.Singleton.ServerClientId, (uint)logType, buffer.Length);
                 }
             }
         }
